+++ extensions/bitwarden/src/components/searchVault/actions/PasteTotpAction.tsx
@@ -1,4 +1,4 @@
-import { Clipboard, Icon, showToast, Toast } from "@raycast/api";
+import { Clipboard, Icon, Keyboard, showToast, Toast } from "@raycast/api";
 import ActionWithReprompt from "~/components/actions/ActionWithReprompt";
 import { useBitwarden } from "~/context/bitwarden";
 import { useSelectedVaultItem } from "~/components/searchVault/context/vaultItem";
@@ -6,7 +6,12 @@ import useGetUpdatedVaultItem from "~/components/searchVault/utils/useGetUpdated
 import { captureException } from "~/utils/development";
 import useFrontmostApplicationName from "~/utils/hooks/useFrontmostApplicationName";
 
-function PasteTotpAction() {
+type PasteTotpActionProps = {
+  omitShortcut?: boolean;
+  shortcut?: Keyboard.Shortcut;
+};
+
+function PasteTotpAction({ omitShortcut, shortcut }: PasteTotpActionProps) {
   const bitwarden = useBitwarden();
   const selectedItem = useSelectedVaultItem();
   const getUpdatedVaultItem = useGetUpdatedVaultItem();
@@ -35,10 +40,14 @@ function PasteTotpAction() {
       title={currentApplicationName ? `Paste TOTP into ${currentApplicationName}` : "Paste TOTP"}
       icon={Icon.Window}
       onAction={pasteTotp}
-      shortcut={{
-        macOS: { key: "t", modifiers: ["opt", "shift"] },
-        Windows: { key: "t", modifiers: ["alt", "shift"] },
-      }}
+      shortcut={
+        !omitShortcut
+          ? shortcut ?? {
+              macOS: { key: "t", modifiers: ["opt", "shift"] },
+              Windows: { key: "t", modifiers: ["alt", "shift"] },
+            }
+          : undefined
+      }
       repromptDescription={`Pasting the TOTP of <${selectedItem.name}>`}
     />
   );